feat: formatTime
