+++ packages/achievements/src/widgets/SudokuBoard.tsx
@@ -8,6 +8,7 @@ import {
   loadStringSetting,
   saveStringSetting,
 } from "../utils/localStorage";
+import { formatTime } from "../utils/timeFormatter";
 
 const getPositionIndex = (n: number) => Math.floor(n / 3);
 
@@ -30,19 +31,6 @@ const loadDifficulty = (): Difficulty => {
 
 const saveDifficulty = (difficulty: Difficulty): void => {
   saveStringSetting(STORAGE_KEYS.DIFFICULTY, difficulty);
-};
-
-const formatTime = (seconds: number): string => {
-  const hrs = Math.floor(seconds / 3600);
-  const mins = Math.floor((seconds % 3600) / 60);
-  const secs = seconds % 60;
-
-  if (hrs > 0) {
-    return `${hrs}:${mins.toString().padStart(2, "0")}:${secs
-      .toString()
-      .padStart(2, "0")}`;
-  }
-  return `${mins}:${secs.toString().padStart(2, "0")}`;
 };
 
 export const SudokuBoard = ({}) => {
